# perf: faster generate_hash