--- a/frappe/model/naming.py
+++ b/frappe/model/naming.py
@@ -278,7 +278,7 @@ def make_autoname(key="", doctype="", doc=""):
 	                DE/09/01/00001 where 09 is the year, 01 is the month and 00001 is the series
 	"""
 	if key == "hash":
-		return frappe.generate_hash(doctype, 10)
+		return frappe.generate_hash(length=10)
 
 	series = NamingSeries(key)
 	return series.generate_next_name(doc)