- Use system randomness
- Ignores `txt` parameter, kept for backward compat. 

~8x faster. 


Before:
```
In [1]: %timeit frappe.generate_hash()
7.57 µs ± 43.9 ns per loop (mean ± std. dev. of 7 runs, 100,000 loops each)

```

After:
```
In [1]: %timeit frappe.generate_hash()
974 ns ± 3.61 ns per loop (mean ± std. dev. of 7 runs, 1,000,000 loops each)
```